requirements.txt
PyYAML~=6.0.2

beautifulsoup4~=4.13.4
Markdown~=3.8
requests~=2.32.3
SQLAlchemy~=2.0.40
torch~=2.7.0
numpy~=2.2.5
sentence-transformers~=4.1.0
sqlean.py~=3.47.0;platform_system=="Darwin"

pytest~=8.3.5
python-dotenv~=1.1.0
psycopg2-binary~=2.9.10
python-dateutil~=2.9.0
python-docx~=1.1.2
pymongo~=4.12.0
openpyxl~=3.1.5
pymupdf~=1.25.5
python-pptx~=1.0.2
boto3~=1.38.3
jsonpath-ng~=1.7.0
openai~=1.76.0

